% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_root.R
\name{.is_omicsflow_root}
\alias{.is_omicsflow_root}
\title{Check whether a directory looks like the OmicsFlow project root}
\usage{
.is_omicsflow_root(path)
}
\arguments{
\item{path}{Directory path to check}
}
\value{
Logical
}
\description{
Check whether a directory looks like the OmicsFlow project root
}
\keyword{internal}
