% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{get_batch}
\alias{get_batch}
\title{Map sample IDs to batches}
\usage{
get_batch(sample_ids, metadata = NULL, omics_type = "rna")
}
\arguments{
\item{sample_ids}{Vector of sample IDs/barcodes}

\item{metadata}{Data frame loaded via load_metadata(), or NULL}

\item{omics_type}{Character, one of "rna", "meth"}
}
\value{
Vector of mapped batches
}
\description{
Map sample IDs to batches
}
\keyword{internal}
