% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{get_sample_class}
\alias{get_sample_class}
\title{Map sample IDs to sample classes/types}
\usage{
get_sample_class(sample_ids, metadata = NULL)
}
\arguments{
\item{sample_ids}{Vector of sample IDs/barcodes}

\item{metadata}{Data frame loaded via load_metadata(), or NULL}
}
\value{
Vector of mapped sample classes/types
}
\description{
Map sample IDs to sample classes/types
}
\keyword{internal}
