% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{msg_fail}
\alias{msg_fail}
\title{Print a failure message to the console}
\usage{
msg_fail(text)
}
\arguments{
\item{text}{Message text}
}
\description{
Print a failure message to the console
}
\keyword{internal}
