% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{msg_info}
\alias{msg_info}
\title{Print an info message to the console}
\usage{
msg_info(text)
}
\arguments{
\item{text}{Message text}
}
\description{
Print an info message to the console
}
\keyword{internal}
