% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{msg_ok}
\alias{msg_ok}
\title{Print a success message to the console}
\usage{
msg_ok(text)
}
\arguments{
\item{text}{Message text}
}
\description{
Print a success message to the console
}
\keyword{internal}
