% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{msg_warn}
\alias{msg_warn}
\title{Print a warning message to the console}
\usage{
msg_warn(text)
}
\arguments{
\item{text}{Message text}
}
\description{
Print a warning message to the console
}
\keyword{internal}
