% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{write_validation_report}
\alias{write_validation_report}
\title{Write a vector of formatted status lines to a text file}
\usage{
write_validation_report(lines, path)
}
\arguments{
\item{lines}{Character vector of report lines}

\item{path}{Output file path}
}
\description{
Write a vector of formatted status lines to a text file
}
\keyword{internal}
