% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_root.R
\name{omicsflow_project_root}
\alias{omicsflow_project_root}
\title{Resolve the OmicsFlow project root directory}
\usage{
omicsflow_project_root()
}
\value{
Absolute path to the OmicsFlow project root.
}
\description{
Resolve the OmicsFlow project root directory
}
\keyword{internal}
